chore: update spring boot 3.5

+++ dossierfacile-pdf-generator/src/test/java/fr/dossierfacile/api/pdfgenerator/service/PdfGeneratorServiceImplTest.java
@@ -19,30 +19,32 @@ import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.Optional;
 
-@SpringBootTest
+@SpringBootTest(properties = "brevo.enabled=false")
 @ExtendWith(ApartmentSharingResolver.class)
 class PdfGeneratorServiceImplTest {
 
     @Autowired
     PdfGeneratorService pdfGeneratorService;
 
-    @MockBean
+    @MockitoBean
     ApartmentSharingCommonService apartmentSharingCommonService;
 
-    @MockBean
+    @MockitoBean
     LogAggregator logAggregator;
 
-    @MockBean
+    @MockitoBean
     TenantCommonRepository tenantRepository;
-    @MockBean
+
+    @MockitoBean
     StorageFileRepository storageFileRepository;
-    @MockBean
+
+    @MockitoBean
     ApplicationLogRepository applicationLogRepository;
 
     @Value("${mock.storage.path}")
@@ -73,7 +75,9 @@ class PdfGeneratorServiceImplTest {
 
     @AfterEach
     void tearDown() {
-        file.delete();
+        if (file != null && file.exists()) {
+            //noinspection ResultOfMethodCallIgnored
+            file.delete();
+        }
     }
-
 }
